src/data/requirements.txt
alabaster==0.7.16
altair==5.5.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
attrs==24.2.0
babel==2.16.0
blinker==1.9.0
boto3==1.35.51
botocore==1.35.51
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.0
coverage==7.6.4
cycler==0.12.1
docutils==0.21.2
flake8==7.1.1
fonttools==4.55.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mccabe==0.7.0
mdurl==0.1.2
minio==7.2.10
narwhals==1.17.0
numpy==2.0.2
packaging==24.1
pandas==2.2.3
pendulum==3.0.0
pillow==11.0.0
plotly==5.24.1
protobuf==5.29.1
psycopg2==2.9.10
pyarrow==18.0.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.21.0
pydeck==0.9.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
s3transfer==0.10.3
seaborn==0.13.2
six==1.16.0
smmap==5.0.1
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.36
streamlit==1.41.0
tenacity==9.0.0
time-machine==2.16.0
toml==0.10.2
tomli==2.0.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
zipp==3.20.2
